+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - feat: Moved MangroveJsDeploy from mangrove-strats to this package. Renamed script to EmptyChainDeployer
 - fix: rounding when deducing tick from price for LiquidityProvider
+- feat: Add spread to market
+- feat: Add getBest to semibook
 
 # 2.0.0